Final polish work before instcon
closes RCX-2073
flag=block_editor

test plan:
  - the background color toolbar button is the same everywhere
  - the image block resize icon is a resize icon and not the compass
  - the button block's icon popup has a white background, not grey
  - the button block's select icon toolbar item is now "Select Icon"
     text, not an icon button
  - the text blocks text color popup is titled "Select a Text Color"
    (not Select Button Color)
  - allie gave me all new icons for the Icon block popup


Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/351622
QA-Review: Jacob DeWar <jacob.dewar@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>

+++ ui/shared/wiki/backbone/views/WikiPageEditView.jsx
@@ -299,7 +299,9 @@ export default class WikiPageEditView extends ValidatedFormView {
   destroyEditor() {
     // hack fix for LF-1134
     try {
-      RichContentEditor.destroyRCE(this.$wikiPageBody)
+      if (!window.ENV.BLOCK_EDITOR) {
+        RichContentEditor.destroyRCE(this.$wikiPageBody)
+      }
     } catch (e) {
       console.warn(e)
     } finally {
@@ -449,7 +451,9 @@ export default class WikiPageEditView extends ValidatedFormView {
     // eslint-disable-next-line no-alert
     if (!this.hasUnsavedChanges() || window.confirm(this.unsavedWarning())) {
       this.checkUnsavedOnLeave = false
-      RichContentEditor.closeRCE(this.$wikiPageBody)
+      if (!window.ENV.BLOCK_EDITOR) {
+        RichContentEditor.closeRCE(this.$wikiPageBody)
+      }
       return this.trigger('cancel')
     }
   }
